Create boolean to check which number to store
Boolean variable created to check whether the user's input should
be stored as the first number or the second number. This will be
later switched when an operator is clicked to allow storage of the
second number.
Also moved code to update calculator display in a seperate function
called display().