Add storage and display of operator symbol
Function added to store the operator button pressed and also display
the operator symbol to the user. If the operator symbol is pressed
multiple times, the most recent operator pressed will replace the
previous - as expected from an calculator.
After an operator button is pressed, all subsequent numbers pressed
are stored as the second number in the equation.