Add ability to carryover the result of expression
The result of the expression is now carried over to allow
an operation to be conducted on the result, and so on, as per
basic functionality of a typical calculator.
The result is stored as the first number of the expression,
and other variables reset to allow for operator and second
number to be input.