Prevent typing of numbers after result displayed
Fixed bug which allowed user to type numbers after getting the
result of an expression without typing the next operator first.
This is undesired behaviour which prevented further calculations
from being performed on the current result. A boolean variable to
keep track of whether an operator has been pressed after determin-
ing the result of an expression has been introduced for this pur-
pose.