Add decimal & delete buttons, move clear button
Clear button has been moved to below the display in a new top-
buttons container, which also includes a new delete button. This was
to allow for the clear button to be replaced by a decimal point.
The delete button will be used in future to provide functionality
for the user to backspace the current input.