Prevent multiple decimal points for single number
Better compatibility for handling decimal number inputs added.
Previously the user could type in multiple decimal points for a
single number, which resulted in a NaN output. A bool variable
has been created to track when a decimal is input, which only
resets when the user moves to the next number input.
Additionally, if the user types a decimal point with no subsequent
input, a trailing zero is added to prevent NaN output.