Add id to decimal btn with seperate event listener
So the code for getting the button value does not need to check if
the value is a decimal point every time a number button is pressed,
the decimal point button has a seperate event listener. The same
checks for the decimal point being present in the current number
are present in the new event listener callback function.